requirements.txt
appdirs==1.4.2
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
Django==1.9
django-braces==1.11.0
django-oauth-toolkit==0.10.0
djangorestframework==3.5.4
enum34==1.1.6
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.1.0
oauthlib==1.0.3
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
requests==2.13.0
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
